// src/types/teamMember.types.ts
export interface TeamMember {
  id: number;

  name: string;

  role: string;

  image: string;

  shortDescription: string;

  story: string;

  experience: string;

  whyNeoversity: string;

  goal: string;

  superpower: string;

  canHelp: string;

  future: string;

  github?: string;

  linkedin?: string;
}